Move the JAX-RS security related filters to the jaxrs extension
